requirements.txt
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
pillow==12.1.0
python-dotenv==1.2.1
python-telegram-bot==20.6
reportlab==4.4.7
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8.1
typing_extensions==4.15.0
urllib3==2.6.3
